feat: env-based table name config
